+++ packages/transaction-manager/test/index.test.ts
@@ -10,6 +10,8 @@ import { DataAccessTypes } from '@requestnetwork/types';
 import { TransactionManager } from '../src/index';
 import TransactionCore from '../src/transaction';
 
+import * as TestData from './unit/utils/test-data';
+
 const channelId = 'channelId';
 const extraTopics = ['topic1', 'topic2'];
 const fakeTxHash = '0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
@@ -34,54 +36,88 @@ const fakeMetaDataAccessGetChannelsReturn: DataAccessTypes.IReturnGetChannelsByT
   meta: { transactionsStorageLocation: { [channelId]: ['fakeDataId1', 'fakeDataId2'] } },
   result: { transactions: { [channelId]: [tx, tx2] } },
 };
-
-const fakeDataAccess: DataAccessTypes.IDataAccess = {
-  getChannelsByTopic: chai.spy.returns(fakeMetaDataAccessGetChannelsReturn),
-  getTransactionsByChannelId: chai.spy.returns(fakeMetaDataAccessGetReturn),
-  initialize: chai.spy(),
-  persistTransaction: chai.spy.returns(fakeMetaDataAccessPersistReturn),
-};
+let fakeDataAccess: DataAccessTypes.IDataAccess;
 
 /* tslint:disable:no-unused-expression */
 describe('index', () => {
-  it('can persist a transaction', async () => {
-    const transactionManager = new TransactionManager(fakeDataAccess);
+  beforeEach(() => {
+    fakeDataAccess = {
+      getChannelsByTopic: chai.spy.returns(fakeMetaDataAccessGetChannelsReturn),
+      getTransactionsByChannelId: chai.spy.returns(fakeMetaDataAccessGetReturn),
+      initialize: chai.spy(),
+      persistTransaction: chai.spy.returns(fakeMetaDataAccessPersistReturn),
+    };
+  });
+  describe('persistTransaction', () => {
+    it('can persist a transaction', async () => {
+      const transactionManager = new TransactionManager(fakeDataAccess);
 
-    const ret = await transactionManager.persistTransaction(data, channelId, extraTopics);
+      const ret = await transactionManager.persistTransaction(data, channelId, extraTopics);
 
-    expect(ret.result, 'ret.result is wrong').to.be.deep.equal({});
-    expect(ret.meta, 'ret.meta is wrong').to.be.deep.equal({
-      dataAccessMeta: fakeMetaDataAccessPersistReturn.meta,
+      expect(ret.result, 'ret.result is wrong').to.be.deep.equal({});
+      expect(ret.meta, 'ret.meta is wrong').to.be.deep.equal({
+        dataAccessMeta: fakeMetaDataAccessPersistReturn.meta,
+      });
+      expect(fakeDataAccess.persistTransaction).to.have.been.called.with(
+        TransactionCore.createTransaction(data),
+        extraTopics,
+      );
     });
-    expect(fakeDataAccess.persistTransaction).to.have.been.called.with(
-      TransactionCore.createTransaction(data),
-      extraTopics,
-    );
   });
 
-  it('can get transactions by channel id', async () => {
-    const transactionManager = new TransactionManager(fakeDataAccess);
+  describe('persistEncryptedTransaction', () => {
+    it('can persist an encrypted transaction', async () => {
+      const transactionManager = new TransactionManager(fakeDataAccess);
 
-    const ret = await transactionManager.getTransactionsByChannelId(channelId);
+      const ret = await transactionManager.persistEncryptedTransaction(
+        data,
+        channelId,
+        [
+          TestData.idRaw1.encryptionParams,
+          TestData.idRaw2.encryptionParams,
+          TestData.idRaw3.encryptionParams,
+        ],
+        extraTopics,
+      );
 
-    expect(ret.result, 'ret.result is wrong').to.be.deep.equal(fakeMetaDataAccessGetReturn.result);
-    expect(ret.meta, 'ret.meta is wrong').to.be.deep.equal({
-      dataAccessMeta: fakeMetaDataAccessGetReturn.meta,
+      expect(ret.result, 'ret.result is wrong').to.be.deep.equal({});
+      expect(ret.meta, 'ret.meta is wrong').to.be.deep.equal({
+        dataAccessMeta: fakeMetaDataAccessPersistReturn.meta,
+        encryptionMethod: 'ecies-aes256-cbc',
+      });
+      expect(fakeDataAccess.persistTransaction).to.have.been.called.once();
     });
-    expect(fakeDataAccess.getTransactionsByChannelId).to.have.been.called.with(channelId);
   });
 
-  it('can get channels indexed by topics', async () => {
-    const transactionManager = new TransactionManager(fakeDataAccess);
+  describe('getTransactionsByChannelId', () => {
+    it('can get transactions by channel id', async () => {
+      const transactionManager = new TransactionManager(fakeDataAccess);
 
-    const ret = await transactionManager.getChannelsByTopic(extraTopics[0]);
+      const ret = await transactionManager.getTransactionsByChannelId(channelId);
 
-    expect(ret.result, 'ret.result is wrong').to.be.deep.equal(
-      fakeMetaDataAccessGetChannelsReturn.result,
-    );
-    expect(ret.meta, 'ret.meta is wrong').to.be.deep.equal({
-      dataAccessMeta: fakeMetaDataAccessGetChannelsReturn.meta,
+      expect(ret.result, 'ret.result is wrong').to.be.deep.equal(
+        fakeMetaDataAccessGetReturn.result,
+      );
+      expect(ret.meta, 'ret.meta is wrong').to.be.deep.equal({
+        dataAccessMeta: fakeMetaDataAccessGetReturn.meta,
+      });
+      expect(fakeDataAccess.getTransactionsByChannelId).to.have.been.called.with(channelId);
     });
-    expect(fakeDataAccess.getChannelsByTopic).to.have.been.called.with(extraTopics[0]);
+  });
+
+  describe('getChannelsByTopic', () => {
+    it('can get channels indexed by topics', async () => {
+      const transactionManager = new TransactionManager(fakeDataAccess);
+
+      const ret = await transactionManager.getChannelsByTopic(extraTopics[0]);
+
+      expect(ret.result, 'ret.result is wrong').to.be.deep.equal(
+        fakeMetaDataAccessGetChannelsReturn.result,
+      );
+      expect(ret.meta, 'ret.meta is wrong').to.be.deep.equal({
+        dataAccessMeta: fakeMetaDataAccessGetChannelsReturn.meta,
+      });
+      expect(fakeDataAccess.getChannelsByTopic).to.have.been.called.with(extraTopics[0]);
+    });
   });
 });