+++ packages/types/src/index.ts
@@ -1,5 +1,7 @@
 import * as AdvancedLogicTypes from './advanced-logic-types';
 import * as DataAccessTypes from './data-access-types';
+import * as DecryptionProviderTypes from './decryption-provider-types';
+import * as EncryptionTypes from './encryption-types';
 import * as ExtensionTypes from './extension-types';
 import * as IdentityTypes from './identity-types';
 import * as LogTypes from './logger-types';
@@ -13,6 +15,8 @@ import * as TransactionTypes from './transaction-types';
 export {
   AdvancedLogicTypes,
   LogTypes,
+  DecryptionProviderTypes,
+  EncryptionTypes,
   ExtensionTypes,
   RequestLogicTypes,
   DataAccessTypes,