+++ packages/types/src/transaction-types.ts
@@ -1,8 +1,16 @@
+import * as Encryption from './encryption-types';
+
 /** Transaction Manager interface */
 export interface ITransactionManager {
   persistTransaction: (
-    transactionData: string,
+    transactionData: ITransactionData,
     channelId: string,
+    topics?: string[],
+  ) => Promise<IReturnPersistTransaction>;
+  persistEncryptedTransaction: (
+    transactionData: ITransactionData,
+    channelId: string,
+    encryptionParams: Encryption.IEncryptionParameters[],
     topics?: string[],
   ) => Promise<IReturnPersistTransaction>;
   getTransactionsByChannelId: (
@@ -27,6 +35,8 @@ export interface IReturnPersistTransaction {
   meta: {
     /** meta-data from the layer below */
     dataAccessMeta?: any;
+    /** encryption method used if transaction encrypted */
+    encryptionMethod?: string;
   };
   /** result of the execution */
   result: {};
@@ -57,6 +67,12 @@ export interface IReturnGetTransactionsByChannels {
 /** Transaction */
 export interface ITransaction {
   data: ITransactionData;
+  /** Hash of the data before encryption */
+  hash?: string;
+  /** Symmetric key encrypted with asymmetric key from the parties keys, indexed by the hash of their identities */
+  keys?: { [key: string]: string };
+  /** Encryption method */
+  encryptionMethod?: string;
 }
 
 /** Transaction confirmed */