+++ packages/utils/test/crypto.test.ts
@@ -86,4 +86,9 @@ describe('Utils.crypto utils', () => {
     expect(crypto.normalize(null)).to.be.equal('null');
     expect(crypto.normalize(undefined)).to.be.equal('undefined');
   });
+
+  it('can generate32BufferKey()', async () => {
+    const randomKey = await crypto.generate32BufferKey();
+    expect(Buffer.from(randomKey, 'base64').length, 'random32Bytes() error').to.be.equal(32);
+  });
 });